FRONT-95 moved the code into a reusable component ImageToggle
