+++ src/components/LoginPage/_Login.scss
@@ -113,6 +113,7 @@
     right: 0;
     top: 0;
     margin-top: 38px;
+    margin-right: 600px;
     width: 65px;
     cursor: pointer;
     
@@ -165,6 +166,7 @@
 
     .show-pass-btn-login {
         margin-top: 30px;
+        margin-right: 0px;
         width: 55px;
     }
