+++ src/components/SignupPage/_SignUp.scss
@@ -46,6 +46,7 @@
     right: 0;
     top: 0;
     margin-top: 57px;
+    margin-right: 30px;
     width: 65px;
     cursor: pointer;
 }
@@ -199,6 +200,7 @@
       
     .show-pass-btn {
         margin-top: 50px;
+        margin-right: 0px;
         width: 55px;
     }
